FCOM-221, Removed empty line between message and signature. Added Definitions.cs. Use same test keys as other Signature tasks
