+++ Frends.Community.PgpClearTextSignature.Tests/TestData/dontuse-sec.asc
@@ -1,0 +1,59 @@
+-----BEGIN PGP PRIVATE KEY BLOCK-----
+
+lQPGBF0kPugBCADJxt+oHbwPy8WIoGzrCvDwA3B2y1vx+W++JSRoAHMxk9rbVDOa
+aXqX0IWW95zuUEOWhrOg3PvHPqxkLYd3CjNhcvxjlu0xukPyMkGX+rGWHnzljwM4
+Cbn/7Id0gCtZ7qqPCpiMBSq/VEDtqzuIjHZ94zT+FAqL8EN69DZSBFeK043iQBY3
+Em/Sq85SXfr7V5A0j+RE+PogcUOyDhXr3esimBD8cludza0XEnaN2RNjQtudNgUX
+W2zLwV7Lv1FgP0XBu8cI6ZH/OfhXuFitvW9ciyXQtgfdM5TMRhX5lbsIJASVVG2e
+KdUpQwkV0bJG6/HB2kS20A1yRsnuGlSfDnsRABEBAAH+BwMCFNWylIpH0ZfKFxi5
+o7MZko54iS7PHwEOh9EUn5Ag3b/pKNeS27OVawnX7APKKi8Xvu4BpYOhrNG60zHF
+VGp5r8oLEMMJ346qF/uLNhbkafBBMoVlqH2ru0MnXhCmpM7i7G7R35mLhir8F6Ps
+WkOzyI+5KVA2XeGyXnec1I8lbOzQxFeuye8atquUAA1a4dGgBWHZxKm9gQL+e/lz
+FjcEqMQNMp9fVnaKcskHUgtbpbOxx5/g/XJPEeUMVcp+Z1RMhwbKhGaOYEFZUSy+
+yPXyE+Lp8DfLos/rGe+VWwT9EhmL4JNVAwLGztDXDw8HCahNT2+AXeAeKGqCLf4J
+YTInWgJdTnJ3lVOd3U22STUgCACc5z1/Dp3m0v2VRgmjOOSi5nTi3DF3KbZ8C2Ev
+AFnzvAdpw3/QcmwOFt0oSu00xIZB3NfdGoK2zayGf74uygMmssfM7kEtCbcizEhD
+sxyceYkKR27b1peIdCwuanlP37K+L8pMKMs9Zx+CaPDNoh7oj9pQS8obxk94R86x
+xNQCQjATE0sc63gVK3OKgNwYpiX6JCAIt+9HTJfbnc6upgPcZaGm7frby6b00BLN
+G8Za9mbALbhk33NxhHSbn32FYn2nYGsDJewhNiLP5ijRrC63L08VdruRSi5XXO/e
+m4GOg7yfUx4v8FDQVUoJlzS3S5m5/6MLghZSrhoK0/QGOVzW2CJFah3GR63rdd3h
+bTZuCbeLOUbnB+wr1CdvYunVMcPppaA1qsNVnc7srk2fTFhMQIGang/D9Fwoz92a
+Caaz5c97js7Lj+C7Kwtv46//l7R3LRsG/dJZ8n8Rw24k3NqwLzfp6bJYmbkCXtBr
+jvBzjoP1XN7GAvAcVvyEfwwtL+TeY61061NuNqfBddI7EdA2rQDxBDxm7W34b0e4
+6uW092FOlO3YtBZ0ZXN0aSA8dGVzdGlAdGVzdGkuZmk+iQFOBBMBCAA4FiEEZYaD
+nkCrS7gaSE+AQmrabh8osUkFAl0kPugCGwMFCwkIBwIGFQoJCAsCBBYCAwECHgEC
+F4AACgkQQmrabh8osUmJMQf/TcWTIwv6ot7A2dmB8VDFFxjNWFmuWuUYcB80In9K
+lsRJ6s41RtaA2LkYyo+43Rdlww4SXnOOK+jsvBWzHQ3P67fE/rl3NROpTQLVn9AN
+Y8+/jrlnJHCMP3YvKnTlKVO2js/8cceqi8gFp2L0MfGP9p0ONrJhVcGRbF/K4oEL
+dSVvCN3mLFpY2NzsGlXCEfswEOM3gwKTDZZA5GCoxJ5hJGZeksC68I+i21pmV7nj
+FZZcOfnxqqXTjO/GFCYVS65gw6SWhqkyXjSdQ/TBlno7SvG0vfuHa0VnbVEV5Y60
+ssW6WRdWUPO9WFNMbgDxeqI+KGPyRMxYrvNqnMFItDzVAp0DxgRdJD7oAQgAwU76
+mwfoo10ZVRu4shzHVPG6SAbHEnpGaxobRyQis3Iz+Kjb0sGsXG1F9JRVmh+7S4nv
+b/eYlomhNZd7OyKv06OvbEPoNyaLVZBp6S615f/rqEwCB1i9Yi5EbhySh7w/oAq5
+0ZPr8jFzF6l21D9NV7NY8UeL7uG6ecM2hXU4BD/cu/+O1RBhpw6gWs/baQBkA2GI
+9w/iY0ZgYBlFn4UIJDz/R2MuKXlAYblvvk4L1vIPgWFe6PNdITxCcBnV+GgctIqX
+a9UBMyQVs5knCtLorp+6/tZpnPdJ7M5qB2cxs5o/3D9pCK3bWbb1jb0V3hohiTFp
+I2BVxqyPqE4GycRPxwARAQAB/gcDAnli3YLILOecysSS/aBq0O66ueHmSFRL606g
+lJG08LOCuZQvZSR79lI0sMH2mj7CUSghK2jXXLAQulBhAuuw0vOuvA+FDNglepW+
+bPtZnt6gVm5ZDc+2EzzhI4K+sG2tOSoRBcuqPwxkgm/PaXoUkHK+efAlEal+Ir1o
+pOmSC/LB+86HS8Bitx1LhClkYGcnJzaBMQif7OG/mReu7cpD8c/ihsAI+NmvgEtw
+oaAO2G8ydBmXyWyA1hUu+PdAKywiYkwKmnWTGIrQlH3z5LA7g/b51QExO0wqr5/M
+pHRPSSV3ycpTYbTMOiQ9cP3gh572yzASgotDho1JiWVOXDgmif7eS/DRITmzoaOj
+ny6IAmpPH97sXSLGO6ZBA1YY6eSBpfwgIVVT4U8Zwk4Hrk0ihqahK3Tlu8aETxo4
+dtj952FbTk240upVgA7ISjE7Ba62Dvdx+KBpIWZqwIdX/4JMkIqx2ReTd+/QbnEL
+KNnnuq5ud5XJ9TOoncxVKlcV1D/P5yajHmQuDQZhLlxfs+Sy9G+8k7HcqOyH0LiC
+9sZi5EHQjVXiffw0B0Uv1vOFf+CwBbfXb2P3FBCojkkn0N73FF+oxxOfUb6/F+a2
+KUnCpLSLKWyK+lUPLa9uI7kOwRjhwBGIFjRLpPEVzZmLXuZbU2RTor5rueIMoEO/
+HUhl4Mcj9QhXhyUmBFAhlPaCfPDbhtM8wjYYCzMq2XNXaxKd7Iwu8oXN94Ju9xh8
+sFcnJLyXZZygPmFpAVbJlwBrtShM7UHAixVcfjrq9QB00OInP3AuKPBIhd8hshM9
+amALUJtnJgkM/32jx68MuYnMBgBtn/rw63eMViBARkBr7PK+N89f0KYUKMic0+BW
+EdposC5sJUSa1ByLhrNRra7k6Nmtz1EOtduiPVR6j6t1lehcVeO3b7nt+YkBNgQY
+AQgAIBYhBGWGg55Aq0u4GkhPgEJq2m4fKLFJBQJdJD7oAhsMAAoJEEJq2m4fKLFJ
+kygH/jrbvtfR3r7yJcUM5iW7OzY7fFa44to6E8emp8qLsot4acSJIIHnFyM14ikA
+WoPw2K98JFCWnkoIJZ71l/X2wzibaEMbvmTpgzrs/pA0OQVH49L4xN0S7mh+9VPX
+hH5gopePiNEpsOBU6b6QvCA1Vnx6tqkEZEVykQK+YCfnCfKNOMNk6N98KtYBUvfU
+cWuCf4/K9s+1taS5P659z+5tQvokkTtjGMnD1J0Z0pGgTWWfbx1GHFJYSW7QttN3
+whLCzJCVsbXnJs5Sg+c9mX13jS6xBohvbzfn0mQM24zZMGPmlvsz1nouOVYsEM1d
+wX7b4+R+0Ss9fJLCBsyiTuhuoQM=
+=WEv8
+-----END PGP PRIVATE KEY BLOCK-----
