--- Frends.Community.PgpClearTextSignature.Tests/TestData/sec.asc
@@ -1,57 +1,0 @@
------BEGIN PGP PRIVATE KEY BLOCK-----
-Version: Keybase OpenPGP v1.0.0
-Comment: https://keybase.io/crypto
-
-xcFGBFpPSusBBADEIMH7SKHveu9hi/xsBkUibUiTVTIX9imVgAk3VQaff5WoZNRp
-mMHY0qzR5j4/Ou8eKWYDyKKQVdAfqsIfHoL8cnegosiQ/xyoEe7HpRScSgh1c2xX
-1fItlJ5r74+bqGWmr0c+u2Aug5m5I17NNXb9PVdwY1fgyDk2eeYQPaycNwARAQAB
-/gkDCGU/pJJB4/m8YDLNWH5IvdTw5HGbI3cv07T94lut3lqGxy/dcBJHsoQ6uwy1
-Gaz6NhpnTB+j2oMyPDxWro6TCynePBRgvnrXPpO2PDrkElmMQo1/vdnjEE9rb+y3
-QcTod1O1hLqMCB0Sp2nP6ktM7smH7tHpfDgh5UlMXO3Z0yJi8f9eivM8W1gXlRWJ
-y+FVLVQO5CKPm8R6E0jXLGm5wuqQvCMv/dNAMYuO3YFk10c0NIVLx+5IJpkbbu91
-dOlIqHRnXZhjA18KpWkSjbIdk6FVowQ9X2Qd/B2c59q0Wk3s4UpONvmuRGteeRc7
-9jZxEBlzjU8+o8ieUPw+HifYE3nmLPxIDDTn2Z+you7rCNg+1GeMwCPhB4Kf7FAU
-ID+M/T+UAiblr6SVjoQeayyX1JwGvwNC9Kacg3bwGpHgnDaiw0hzMW0xRhhhuSbm
-/b5SfchoTfCwZ6tZdfulO5TTZZ0n8JStHOShbBSka4DGj8KrtBUUlOXNG1Rlc3Qg
-S2V5IDx0ZXN0QGV4YW1wbGUuY29tPsKtBBMBCgAXBQJaT0rrAhsvAwsJBwMVCggC
-HgECF4AACgkQzoB5W4N7pN5CpwQAqtKaLBbvl4Wm4jYGfpoXxr3ic8OryfRtzrXN
-gAKEgv8cQuNOwd9BRmVYvdFoJwWgWn/YarDQvOHPCO6iYhkG8bLMYMvPxiPfjuiJ
-aN1w2UYByg1fCcK7HXDVaFW8FpyNoxB7X9hPcucDiUud4KT86zuXjgcvTCFQWgiK
-85tK3HrHwUYEWk9K6wEEANC2/avaEzhP8rGKHvHmOTdAqyuNWe/Yjfosyu8TGX6T
-tYi3FWGdeCOxztV8ZljxbM6/KVOV5oUZ3VG6ignCzKRK+fJum1WH+rjpD7tuPuWV
-K4BE2zlBt1qn85s9pSFiDP5WbKZZwO3NoqalF6EhwRKlPinhS7agpPq27VyNe8IB
-ABEBAAH+CQMIjcLivvhToYtgwyf7kJcjbOeGfcKrCSsWou/OngiiHHd8i7EbQQmb
-obJIaDq9UHT7e7h5LJvJsNXoDKUjGlfJ7MweUP5WNlN0IxN1iwo6p8coLsAIgYUw
-bQhxZddRVBGrLg46C1uVJErEorNEVo4FIWwR3pvOIvG7P03pvFrFhvMIfeXUItFw
-77+6+yVDu/mEdtu5brx989U7X+bpQ1FU6PBRVFM9caZoP1zQnGugKqWbyVa0D6rh
-URhBo36sDx0omzQz07VqGzO9sMkHM6WI3+E/H1rH5Wrocja+GlxZReuJz0wtyLm3
-ptwZO2IMGtS8+r2W0r36xn9An4394aToytHr5GVztZZlUZflq6lQ90Q8WRVcKbY2
-6t5EosYub5fD+VUp4pHjAb890nDqxzJtCi2xuTqDiUl50fh40dLDBbmfSEMeRg/Z
-9fK8a/QyjgRsrdhdY/6WCImJUTjSSOLX8oZexVIKaBTf4kA4LOm9gKtlm/GCAMLA
-gwQYAQoADwUCWk9K6wUJDwmcAAIbLgCoCRDOgHlbg3uk3p0gBBkBCgAGBQJaT0rr
-AAoJEPfgQtloRS5slHQD/R/eWB+gIoSIRk2DStj/FF+jdqGmk+MXcJ2tXo3N/CYL
-6tPuXa/Ch9aWkyaTwbxx+Fqc6yZgdQd+Z7Fqad+uZ7eChBnPHAamuYuTDcgSTW+4
-5yUsxF6NtTZeNP50EOlzL0FKpOTUG9qL4azfD7ZmPsYJr0cWP1AeaWbF3jPV2opc
-NwoD/RG7lq5B6EC6KFS6zaBf0mw2Qp8Wm6oe0EKLr8IPFlPF50Y/loDxpW6L9qnV
-ISYXqykKcjUn6zOBpFoJAj2Bt33uCH75MwrYfJCA2OI7gBPiJzVntvGbRYdEoxij
-wpBn1Kcf6dpZvyhUT1UAGiEwGz+LRzfkWS2DHDZWAv504Ab+x8FGBFpPSusBBADy
-fXAW9dvhZ4NhNCxj/ln22BJwsslXFntL9UuvA40xT3r6de7ZeqJQEyGdXWwuBBrM
-OW5+MAt6Gw964l6SLGn0M8bMuvzSolrejScvjCYgKsODqnjtjYWRGUsq4r5lyHvT
-zsT1CfsZGS91PFED2wHrQRAK8K5CL1s5xdK0pO3ZHQARAQAB/gkDCHeZsDo0zzg2
-YB7Ktfwajf/jVmB6mbHJWD86rbR3P+sruhj2CWu6ZabU/qNvqlmch1jLeeQ4svVt
-bTkFRzgmBsO4j+utHtvgm5VRNGObjpTJDSL6VriDs4lrDauO9diWo+xyS8CBnL+1
-VGsg94x+76kX7VE71DI+40LSpUHXGGiL4K47oSN93qBy8MMWlK2VPTG3swy91eWV
-R5n5AvOreXgrY/a2eOh07KbkIrhCkA4IKBLO7WPoV6gsrUmKbF3o65TbLk65JzJw
-0PxoC2EKMAw3ORJShnT9YEM5JKrz2A6wATwyHVUMJmz/LMZXRiSelIC6Dso2YMaT
-3SiL6IooqNO+Ams6BACBFmgJNxxvICr9VCDe5F33qkWgcUWRP0yp9bXCibiyau3k
-RIlIxBYkBZAzc3QW4YXmGBKxCDfz7JeMeAIk/RLRb5mc9H4Ux93+E9rDeul5YnZ+
-yW7RXo1ehk3aluW2vMZyuAYquOhu9VDT14ojCK/CwIMEGAEKAA8FAlpPSusFCQ8J
-nAACGy4AqAkQzoB5W4N7pN6dIAQZAQoABgUCWk9K6wAKCRB6gD/0jASUHcK0BADL
-WExUCvwrDNzQeG/G1CtYRRP5yNeV+ObYuNxZZancCLrdQM4Illu70s0DPmh7WW15
-3cKRZWqumKI5x6uwf10M2C796z74QBdZ+PLjFUrGfecw98KgiVl2/C37+RTrZpRx
-YbDPFiMfgyShi7Ka/fMXRWn+2vAZ6o5o/tJUZiS7+NCZA/wNyZpcIXwafaezW7Qn
-f289vGBHfJCL97A7bSgRn1WLgvsPOdXbP4ALVotOZ5I/X2fB6JYpA8yf1Ja+WgbE
-D/xWrSw86e/somWxKkbBxYyndipMoa0hCef6XB0TurnThoPtjbNCYiTADQ0qaY9+
-5XXtXL6tQY0CgkQXXfuhh5PR4Q==
-=dGMf
------END PGP PRIVATE KEY BLOCK-----
