+++ Frends.Community.PgpClearTextSignature/Frends.Community.PgpClearTextSignature.csproj
@@ -53,7 +53,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PgpClearTextSignature.cs" />
+    <Compile Include="Definitions.cs" />
+    <Compile Include="PgpClearTextSignatureTask.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
