+++ nuspec/Frends.Community.PgpClearTextSignature.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
 	<metadata>
 		<id>Frends.Community.PgpClearTextSignature</id>
-		<version>1.0.0</version>
+		<version>1.0.1</version>
 		<title>Frends Community PgpClearTextSignature</title>
 		<authors>HiQ Finland</authors>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
